Move LocalResponse to CssValidator namespace
