+++ lib/headhunter.rb
@@ -1,6 +1,5 @@
 require 'headhunter/css_hunter'
 require 'headhunter/css_validator'
-require 'headhunter/local_response'
 require 'headhunter/html_validator'
 require 'headhunter/rails'
 require 'headhunter/runner'
